fix: actionsに`auth`の処理追加・prismaロジック分割
