+++ actions/circle/add-activity.ts
@@ -1,6 +1,6 @@
-// /app/(circle)/[circleId]/activities/actions.ts
 "use server"
 
+import { auth } from "@/auth"
 import { createActivity } from "@/data/activity"
 import { getMemberByCircleId } from "@/data/circle"
 import { getUserById } from "@/data/user"
@@ -11,6 +11,14 @@ export const addActivityAction = async (
   circleId: string,
   userId: string,
 ) => {
+  // 認証情報を取得
+  const session = await auth()
+
+  // 認証されたユーザーIDとリクエストのuserIdが一致しているか確認
+  if (!session?.user || session.user.id !== userId) {
+    return { success: false, error: "権限がありません。" }
+  }
+
   // メンバー情報を取得
   const members = await getMemberByCircleId(circleId)
 
@@ -18,11 +26,11 @@ export const addActivityAction = async (
   const isMember = members?.some((member) => member.id === userId)
 
   if (!isMember) {
-    throw new Error("権限がありません。")
+    return { success: false, error: "権限がありません。" }
   }
   const user = await getUserById(userId)
   if (!user) {
-    throw new Error("ユーザーが存在しません。")
+    return { success: false, error: "ユーザーが存在しません。" }
   }
   try {
     // Prismaのロジックを呼び出し
