+++ actions/circle/delete-circle.ts
@@ -1,9 +1,20 @@
 "use server"
 
+import { auth } from "@/auth"
 import { deleteCircle, isUserAdmin } from "@/data/circle"
 
 // サーバーアクション：サークル削除
 export const removeCircle = async (circleId: string, userId: string) => {
+  // 認証情報を取得
+  const session = await auth()
+
+  // 認証されたユーザーIDとリクエストのuserIdが一致しているか確認
+  if (!session?.user || session.user.id !== userId) {
+    return {
+      success: false,
+      message: "権限がありません。",
+    }
+  }
   try {
     const currentUser = await isUserAdmin(userId, circleId)
     if (!currentUser || currentUser.roleId !== 0) {
