+++ actions/circle/fetch-activity.ts
@@ -1,10 +1,21 @@
 "use server"
 
+import { auth } from "@/auth"
 import { getActivitiesByMonth } from "@/data/activity"
 
 // サーバーアクション：指定月のイベントを取得
 export const fetchActivitiesByMonth = async (date: Date, circleId: string) => {
   try {
+    // 認証情報を取得
+    const session = await auth()
+
+    // 認証されたユーザーIDとリクエストのuserIdが一致しているか確認
+    if (!session?.user) {
+      return {
+        success: false,
+        message: "権限がありません。",
+      }
+    }
     const year = date.getFullYear()
     const month = date.getMonth() + 1 // 月は1月が0から始まるため+1
