+++ app/circles/[circle_id]/[tab_key]/page.tsx
@@ -1,7 +1,7 @@
+import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { getMembershipRequests } from "@/actions/circle/membership-request"
 import { auth } from "@/auth"
 import { CircleDetailPage } from "@/components/layouts/circle-detail-page"
-import { getCircleById, getCircles } from "@/data/circle"
 
 interface Props {
   params: {
