+++ app/page.tsx
@@ -19,9 +19,9 @@ import {
   VStack,
 } from "@yamada-ui/react"
 import Link from "next/link"
+import { getCirclesByUserId } from "@/actions/circle/fetch-circle"
 import { auth } from "@/auth"
 import { CircleCard } from "@/components/data-display/circle-card"
-import { getCirclesByUserId } from "@/data/circle"
 import { getUserById } from "@/data/user"
 
 const notificationMockData = [
