+++ components/data-display/member-card.tsx
@@ -24,7 +24,7 @@ import {
   Button,
 } from "@yamada-ui/react"
 import { useState, type FC } from "react"
-import type { getCircleById } from "@/data/circle"
+import type { getCircleById } from "@/actions/circle/fetch-circle"
 
 interface MemberCard {
   member: NonNullable<
