+++ components/forms/circle-form.tsx
@@ -34,8 +34,8 @@ import { useState } from "react"
 import { Controller, useForm } from "react-hook-form"
 import { CreateCircle } from "@/actions/circle/create-circle"
 import { removeCircle } from "@/actions/circle/delete-circle"
+import type { getCircleById } from "@/actions/circle/fetch-circle"
 import { UpdateCircle } from "@/actions/circle/update-circle"
-import { type getCircleById } from "@/data/circle"
 import { BackCircleSchema, FrontCircleSchema } from "@/schema/circle"
 import type { FrontCircleForm } from "@/schema/circle"
