+++ data/membership.ts
@@ -1,4 +1,3 @@
-"use server"
 import { db } from "@/utils/db"
 
 export const createMembershipRequest = async (
